requirements.txt
streamlit
scikit-learn
nltk
ntscraper
numpy
pandas
openai
python-dotenv
matplotlib
seaborn
fpdf 